--- maeilmail-admin/src/main/java/maeilmail/admin/AdminNoticeTestRequest.java
@@ -1,4 +1,0 @@
-package maeilmail.admin;
-
-record AdminNoticeTestRequest(String target) {
-}
