+++ maeilmail-admin/src/main/java/maeilmail/admin/AdminReportScheduler.java
@@ -4,6 +4,9 @@ import java.time.LocalDate;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import maeilmail.admin.domain.Admin;
+import maeilmail.admin.domain.AdminRepository;
+import maeilmail.admin.view.AdminReportView;
 import maeilmail.mail.MailMessage;
 import maeilmail.mail.MailSender;
 import maeilmail.statistics.DailySendReport;
