--- maeilmail-admin/src/main/java/maeilmail/admin/AdminRepository.java
@@ -1,6 +1,0 @@
-package maeilmail.admin;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface AdminRepository extends JpaRepository<Admin, Long> {
-}
