+++ maeilmail-admin/src/main/java/maeilmail/admin/api/AdminAuthInterceptor.java
@@ -1,4 +1,4 @@
-package maeilmail.admin;
+package maeilmail.admin.api;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
