+++ maeilmail-admin/src/main/java/maeilmail/admin/application/AdminNoticeTestRequest.java
@@ -1,0 +1,4 @@
+package maeilmail.admin.application;
+
+public record AdminNoticeTestRequest(String target) {
+}
