+++ maeilmail-admin/src/main/java/maeilmail/admin/domain/Admin.java
@@ -1,4 +1,4 @@
-package maeilmail.admin;
+package maeilmail.admin.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,10 +6,11 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.Getter;
+import maeilmail.admin.BaseEntity;
 
 @Getter
 @Entity
-class Admin extends BaseEntity {
+public class Admin extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
