+++ maeilmail-admin/src/main/java/maeilmail/admin/domain/AdminRepository.java
@@ -1,0 +1,6 @@
+package maeilmail.admin.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminRepository extends JpaRepository<Admin, Long> {
+}
