+++ maeilmail-app/src/main/java/maeilmail/config/WebMvcConfig.java
@@ -3,7 +3,7 @@ package maeilmail.config;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 
-import maeilmail.admin.AdminAuthInterceptor;
+import maeilmail.admin.api.AdminAuthInterceptor;
 import maeilwiki.auth.MemberIdentityArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
